feat: add journal-api-status script (#4)
* feat: add journal-api-status script

This commit adds a new script `journal-api-status.js` which fetches and compares
semi-automated and automated snippets from eLife Sciences API endpoints. It also
updates the `package.json` file to include a new script command to run
`journal-api-status.js`.

* feat: Improve console output and update README

The console output in `journal-api-status.js` has been updated for better readability.
Also, added instructions in `README.md` on verifying the status of the automated instance of Journal API.

* feat: Improve console output and update README

The console output in `journal-api-status.js` has been updated for better readability.
Also, added instructions in `README.md` on verifying the status of the automated instance of Journal API.

* feat: update snippet collection in journal-api-status.js

This commit updates the snippet collection in journal-api-status.js. It
changes the condition for collecting semi-automated and automated snippets
to gather all snippets instead of just the first 50. Also, it comments out
the PDF availability in automated listings as it's not yet available.

* fix: correct log message formatting in journal-api-status.js